Remove code duplication in InclusionValidator and ExclusionValidator.
Orginal code by @jamescook

Closes #1352

[Rafael Mendonça França + James Cook]
